pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "fedmpq_kd_framework"
version = "0.1.0"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "numpy>=1.23",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
]
